Apicarto module Geoportail de l'urbanisme

Apicarto module Geoportail de l'urbanisme

Usage

get_apicarto_plu(x,
                 ressource = "zone-urba",
                 partition = NULL,
                 timeout = 10)

Arguments

x

An object of clas sf or sfc. If NULL, partition must be filled by partition of PLU.

ressource

A character from this list : "document", "zone-urba", "secteur-cc", "prescription-surf", "prescription-lin", "prescription-pct", "info-surf", "info-lin", "info-pct". See detail for more info.

partition

A character corresponding to PLU partition (can be retrieve using get_apicarto_plu(x, "document", partition = NULL)). If partition is explicitely set, all PLU features are returned and geom is override

timeout

Time to wait between two request. It's useful when get_apicarto_plu() is implemented inside dynamic document. If API doesn't work for some reasons, it will try again (3 times). Be careful, if you're download is longer than timeout, the download will not have time to complete

Value

A object of class sf

Details

"document' :

"zone-urba" :

"secteur-cc" :

"prescription-surf" :

"prescription-lin" :

"prescription-pct" :

"info-surf" :

"info-lin" :

"info-pct" :

Examples

if (FALSE) {
library(tmap)
library(sf)
point <- st_sfc(st_point(c(-0.4950188466302029, 45.428039987269926)), crs = 4326)

# If you know the partition (all PLU features are returned, geom is override)
partition <- "DU_17345"
poly <- get_apicarto_plu(x = NULL, ressource = "zone-urba", partition = partition)
qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)

# If you don't know partition (only intersection between geom and PLU features is returned)
poly <- get_apicarto_plu(x = point, ressource = "zone-urba", partition = NULL)
qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)

# If you wanna find partition
document <- get_apicarto_plu(point, ressource = "document", partition = NULL)
partition <- unique(document$partition)

# Get all prescription : /!\ prescription is different than zone-urba
partition <- "DU_17345"
ressources <- c("prescription-surf", "prescription-lin", "prescription-pct")

# I recommend to use purrr package for loop
library(purrr)
all_prescription <- map(.x = ressources,
                        .f = ~ get_apicarto_plu(point, .x, partition))
}