Create isochron or isodistance from IGN routing API

Create isochron or isodistance from IGN routing API

Usage

get_iso(point,
               time = NULL,
               distance = NULL,
               transport = "Voiture",
               reverse = FALSE,
               smoothing = TRUE,
               holes = TRUE,
               crs = "EPSG:4326")

Arguments

point

Object of class sf or sfc. Needs to be located in France.

time

Duration in minute to create isochron. If time and distance parameters are set, only time will be used.

distance

Distance in meter to create isodistance. If time and distance parameters are set, only time will be used.

transport

There is two type of transport : "Voiture" (car) or " Pieton" (pedestrian).

reverse

A boolean corresponding to the sens of travel. If FALSE, point corresponding to the start.

smoothing

A boolean determining whether the returned shape should be smoothed or not

holes

Some areas are inaccessible. If TRUE they are highlighted

crs

System of coordinate

Value

get_iso() returned a polygon corresponding to all the points reachable in a defined time or distance

Details

Native API parameter can be find at IGN documentation

Examples

if (FALSE) {
library(sf)
library(tmap); tmap_mode("view")

point <- st_sfc(st_point(c(-4.340329361590764, 47.81236909325107)), crs = 4326)
isodistance <- get_iso(point, distance = 1000, transport = "Pieton")

qtm(isodistance)+
qtm(point, dots.col = "red")

}