isochronous/isodistance calculations

Calculates isochrones or isodistances in France from an sf object using the IGN API on the Géoportail platform. The reference data comes from the IGN BD TOPO® database. For further information see IGN documentation.

Usage

get_iso(x,
        value,
        type = "time",
        profile = "pedestrian",
        time_unit = "minute",
        distance_unit = "meter",
        direction = "departure",
        source = "iso",
        constraints = NULL)

get_isodistance(x,
                dist,
                unit = "meter",
                source = "iso",
                profile = "car",
                direction = "departure",
                constraints = NULL)

get_isochrone(x,
              time,
              unit = "minute",
              source = "iso",
              profile = "car",
              direction = "departure",
              constraints = NULL)

Arguments

x

Object of class sf or sfc with POINT geometry. There may be several points in the object. In this case, the output will contain as many polygons as points.

value

numeric; A quantity of time or distance.

type

character; Specifies the type of calculation performed: "time" for isochrone or "distance" for isodistance (isochrone by default).

profile

character; Type of cost used for calculation: "pedestrian" for #' pedestrians and "car" for cars. and "car" for cars ("pedestrian" by default).

time_unit

character; Allows you to specify the unit in which times are expressed in the answer: "hour", "minute" or "second" (minutes by default).

distance_unit

character; Allows you to specify the unit in which distances are expressed in the answer: "meter" or "kilometer" (meter by default).

direction

character; Direction of travel. Either define a "departure" point and obtain the potential arrival points. Or define an "arrival" point and obtain the potential points ("departure" by default).

source

character; This parameter specifies which source will be used for the calculation. Currently, "iso" and "pgr" sources are available (default "iso"). See section SOURCE for further information.

constraints

Used to express constraints on the characteristics to calculate isochrones/isodistances. See section CONSTRAINTS.

dist

numeric; A quantity of time.

unit

see time_unit and distance_unit param.

time

numeric; A quantity of time.

Value

object of class sf with POLYGON geometry

Functions

get_isodistance(): Wrapper function to calculate isodistance from get_iso.

get_isochrone(): Wrapper function to calculate isochrone from get_iso.

SOURCE

Isochrones are calculated using the same resources as for route calculation. PGR" and "ISO" resources are used, namely "bdtopo-iso" and "bdtopo-pgr".

bdtopo-pgr" is based solely on the new engine, but lacks performance on large isochrones. On the other hand, it is functional for small isochrones.

bdtopo-iso" is based on the old services over a certain distance, to solve performance problems. We recommend its use for large isochrones.

PGR resources are resources that use the PGRouting engine to calculate isochrones. ISO resources are more generic. The engine used for calculations varies according to several parameters. At present, the parameter concerned is cost_value, i.e. the requested time or distance.

See also