Download WMS raster layer

Directly download a raster layer from the French National Institute of Geographic and Forestry. To do that, it need a location giving by a shapefile, an apikey and the name of layer. You can find those information from IGN website

Usage

get_wms_raster(shape,
               apikey,
               layer_name,
               resolution = NULL)

Arguments

shape

Object of class sf. Needs to be located in France.

apikey

API key from get_apikeys() or directly from IGN website

layer_name

Name of the layer from get_layers_metadata(apikey, "wms") or directly from IGN website

resolution

Cell size in meter. WMS are limited to 2048x2048 pixels. See detail for more information about resolution.

Value

get_wms_raster return an object of class stars. Depending on the layer, this can be a simple raster (2 dimensions and 1 attribute) or an RGB raster (3 dimensions and 1 attribute).

Details

If the resolution is too high, the maximum is automatically set by dividing height and width of the shape's bbox by 2048 (the maximum number of pixel)

Setting the resolution parameter higher than the base resolution of the layer multiplies the number of pixels without increasing the precision. For example, the download of the BD Alti layer from IGN will be optimal for a resolution of 25m. Look at IGN documentation for more precision on layer's resolution.

See also

Examples

if (FALSE) {
library(sf)
library(stars)

apikey <- get_apikeys()[4]

metadata_table <- get_layers_metadata(apikey, "wms")
layer_name <- metadata_table[1,2]

# shape from the best town in France
shape <- st_polygon(list(matrix(c(-4.373937, 47.79859,
                                 -4.375615, 47.79738,
                                 -4.375147, 47.79683,
                                 -4.373898, 47.79790,
                                 -4.373937, 47.79859),
                                 ncol = 2, byrow = TRUE)))
shape <- st_sfc(shape, crs = st_crs(4326))

# Downloading digital elevation model from IGN
MNT <- get_wms_raster(shape, apikey, layer_name, resolution = 25)

# Verif
plot(MNT)
plot(shape, col = NA, add = TRUE)
}