Download WMS raster layer

Download a raster layer from IGN Web Mapping Services (WMS). To do that, it need a location giving by a shape, an apikey and the name of layer. You can find those information from IGN website or with get_apikeys() and get_layers_metadata().

Usage

get_wms_raster(shape,
               apikey = "altimetrie",
               layer_name = "ELEVATION.ELEVATIONGRIDCOVERAGE.HIGHRES",
               resolution = 5,
               filename = tempfile(fileext = ".tif"),
               crs = 2154,
               overwrite = FALSE,
               version = "1.3.0",
               styles = "",
               interactive = FALSE)

Arguments

shape

Object of class sf. Needs to be located in France.

apikey

API key from get_apikeys() or directly from IGN website.

layer_name

Name of the layer from get_layers_metadata(apikey, "wms") or directly from IGN website.

resolution

Cell size in meter. See detail for more information about resolution.

filename

Either a character string naming a file or a connection open for writing. (ex : "test.tif" or "~/test.tif"). If NULL, layer_name is used. Default drivers is ".tif" but all gdal drivers are supported, see details for more info. To avoid re-downloads, get_wms raster checks that there is not already a raster with that name. If it does, it is imported into R without further downloading if overwrite is set to FALSE.

crs

Numeric, character, or object of class sf or sfc. It is set to EPSG:2154 by default. See sf::st_crs() for more detail.

overwrite

If TRUE, output raster is overwrite.

version

The version of the service used. See detail for more information about version.

styles

The rendering style of the layers. Set to "" by default. See detail for more information about styles.

interactive

If set to TRUE, no need to specify apikey and layer_name, you'll be ask.

Value

get_wms_raster return an object of class SpatRaster from terra package.

Details

Raster tile are limited to 2048x2048 pixels so depending of the shape and the resolution, correct number of tiles to download is calculated. This mean that setting the resolution argument higher than the base resolution of the layer multiplies the number of pixels without increasing the precision. For example, the download of the BD Alti layer from IGN will be optimal for a resolution of 25m.

version and styles arguments are detailed on IGN documentation

Using the crs argument avoids post-reprojection which can be time consuming

All GDAL supported drivers can be found here

See also

Examples

if (FALSE) {
library(sf)
library(tmap)

# shape from the best town in France
penmarch <- read_sf(system.file("extdata/penmarch.shp", package = "happign"))

# For quick testing, use interactive = TRUE
raster <- get_wms_raster(shape = penmarch, interactive = TRUE)

# For specific use, choose apikey with get_apikey() and layer_name with get_layers_metadata()
apikey <- get_apikeys()[4]  # altimetrie
metadata_table <- get_layers_metadata(apikey, "wms") # all layers for altimetrie wms
layer_name <- as.character(metadata_table[2,1]) # ELEVATION.ELEVATIONGRIDCOVERAGE

# Downloading digital elevation model from IGN
mnt <- get_wms_raster(penmarch, apikey, layer_name, resolution = 25)

# Preparing raster for plotting
mnt[mnt < 0] <- NA # remove negative values in case of singularity
names(mnt) <- "Elevation [m]" # Rename raster ie the title legend

# Plotting result
tm_shape(mnt)+
   tm_raster(legend.show = FALSE)+
tm_shape(penmarch)+
   tm_borders(col = "blue", lwd  = 3)
}